requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
black==23.10.1
click==8.1.7
decorator==5.1.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==2.0.0
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.12.0
typing_extensions==4.8.0
wcwidth==0.2.8
